backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pymysql
pymssql
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain-ollama
chromadb
sentence-transformers==3.0.0
pypdf==4.2.0
python-docx==1.1.0
docx2txt==0.9
pandas==2.2.2
openpyxl==3.1.2
celery==5.4.0
redis==5.0.4
httpx==0.27.0
python-dotenv==1.0.1
streamlit==1.35.0
requests==2.32.2
